console: render environment-not-ready flow in the standard layout
The environment-not-ready flow used a stripped custom layout with no
navigation, so users without an enabled environment could not see or
reach account-scoped pages like App Passwords, License, or Usage &
Billing. A blocked or trial-expired organization, by contrast, already
rendered the full BaseLayout with those links visible, making the two
states inconsistent.

Render EnvironmentNotReadyRoutes inside BaseLayout instead. The nav
already handles the no-environment state: region-scoped items hide via
HideIfEnvironmentDisabled while account-scoped Admin items carry
forceShow, so new users now get the same chrome as everyone else. The
logo links back to the enable-region flow while no environment is
ready.

The welcome dialog is suppressed in this flow via a new BaseLayout
prop. It pops the moment a region becomes healthy and would cover the
flow's own region-ready affordances, blocking the tutorial's "Open
console" button (caught by the e2e suite).

The region-ready toast is preserved in a headless RegionReadyToast
component. Its unmount cleanup reads the unstable toast reference
through a ref instead of an eslint-disable, which the react-compiler
lint rule no longer accepts. The custom layout file is deleted.

Adds tests asserting the Admin nav group renders on the enable-region
page with no enabled environment while region-scoped items stay
hidden, and that the welcome dialog stays suppressed once a region is
healthy. The test provider wrapper now nests ToastProvider inside the
router to match the app, so toasts can render router Links.
