+++ console/src/layouts/BaseLayout.tsx
@@ -54,6 +54,12 @@ import { FIXED_TOP_BAR_Z_INDEX, MAIN_CONTENT_Z_INDEX } from "./zIndex";
 export interface BaseLayoutProps {
   children?: React.ReactNode;
   containerProps?: FlexProps;
+  /**
+   * Suppresses the welcome dialog that pops when a region first becomes
+   * healthy. Used by flows with their own region-ready affordances, so the
+   * dialog doesn't cover them.
+   */
+  hideWelcomeDialog?: boolean;
   navBarOverride?: React.FunctionComponent;
   sectionNav?: React.ReactNode;
 }
@@ -98,7 +104,7 @@ export const BaseLayout = (props: BaseLayoutProps) => {
       data-testid="page-layout"
       {...props.containerProps}
     >
-      <WelcomeDialog />
+      {!props.hideWelcomeDialog && <WelcomeDialog />}
       <MfaAlert />
       <ImpersonationAlert />
       <Flex
